Switch from font-manager to system-font-families
Now it will only display TrueType and OpenType fonts, which actually
works, while it used to be that it would also show some fonts (bitmap?)
that doesn't work (which LibreOffice doesn't display), thanks to
fontconfig on Linux.

Furthermore, since font-manager is a native Node module it means that
compatibility with newer Node.js versions can be a problem (it fails to
build on version 12 for example), not to mention that we have to rebuild
it for Electron's specific Node.js version as well (this was why
electron-rebuild was added into the yarn build script previously).